Updated Image storage
Now, a new image will not get stored immediately until submit is called, reducing the number of useless images we have in storage

+++ ripples/src/components/BeneficiaryPages/BeneficiarySettingPageFolder/BeneficiaryUpdateProfile.js
@@ -40,18 +40,15 @@ export default function BeneficiaryUpdateProfile() {
     "Please select an Image for your Profile"
   );
   const [fileUrl, setFileUrl] = useState(null);
+  const [file, setFile] = useState(null);
 
   const onFileChange = async (e) => {
-    const file = e.target.files[0];
-    const storageRef = storage.ref();
-    const fileUID = uuidv4();
-    const fileRef = storageRef.child(fileUID);
-    await fileRef.put(file);
-    setDisplayImage(file.name);
-    setFileUrl(await fileRef.getDownloadURL());
+    const currentFile = e.target.files[0];
+    setFile(currentFile);
+    setDisplayImage(currentFile.name);
   };
 
-  function handleSubmit(e) {
+  async function handleSubmit(e) {
     e.preventDefault();
     if (passwordRef.current.value !== passwordConfirmRef.current.value) {
       return setError("Passwords do not match");
@@ -94,11 +91,18 @@ export default function BeneficiaryUpdateProfile() {
     if (dobRef.current.value) {
       updates.push(updateDob(dobRef.current.value, currentUser.uid));
     }
-    if (fileUrl) {
+
+    if (file) {
+      const oldFileUrlRef = dbUser.fileUrl;
+      var oldRef = storage.refFromURL(oldFileUrlRef);
+      const storageRef = storage.ref();
+      const fileUID = uuidv4();
+      const fileRef = storageRef.child(fileUID);
+      await fileRef.put(file);
+      setFileUrl(await fileRef.getDownloadURL());
       updates.push(updateProfileFileUrl(fileUrl, currentUser.uid));
     }
-    const oldFileUrlRef = dbUser.fileUrl;
-    var oldRef = storage.refFromURL(oldFileUrlRef);
+
     Promise.all(updates)
       .then(() => {
         history.push("/");
@@ -109,16 +113,20 @@ export default function BeneficiaryUpdateProfile() {
       .finally(() => {
         setLoading(false);
       });
-    oldRef
-      .delete()
-      .then(() => {
-        // File deleted successfully
-        console.log("successfully deleted " + oldFileUrlRef);
-      })
-      .catch((error) => {
-        // Uh-oh, an error occurred!
-        console.log("unable to delete " + oldFileUrlRef);
-      });
+
+    // Delete the file
+    if (oldRef) {
+      oldRef
+        .delete()
+        .then(() => {
+          // File deleted successfully
+          console.log("successfully deleted " + oldRef);
+        })
+        .catch((error) => {
+          // Uh-oh, an error occurred!
+          console.log("unable to delete " + oldRef);
+        });
+    }
   }
 
   function openModal() {
