+++ ripples/src/components/VolunteerPages/VolunteerSettingsPageFolder/VolunteerUpdateProfile.js
@@ -40,20 +40,17 @@ export default function VolunteerUpdateProfile() {
     "Please select an Image for your Profile"
   );
   const [fileUrl, setFileUrl] = useState(null);
+  const [file, setFile] = useState(null);
 
   const onFileChange = async (e) => {
-    const file = e.target.files[0];
-    const storageRef = storage.ref();
-    const fileUID = uuidv4();
-    const fileRef = storageRef.child(fileUID);
-    await fileRef.put(file);
-    setDisplayImage(file.name);
-    setFileUrl(await fileRef.getDownloadURL());
+    const currentFile = e.target.files[0];
+    setFile(currentFile);
+    setDisplayImage(currentFile.name);
   };
 
   console.log(dbUser);
 
-  function handleSubmit(e) {
+  async function handleSubmit(e) {
     e.preventDefault();
     if (passwordRef.current.value !== passwordConfirmRef.current.value) {
       return setError("Passwords do not match");
@@ -106,12 +103,16 @@ export default function VolunteerUpdateProfile() {
       updates.push(updateDob(dobRef.current.value, currentUser.uid));
     }
 
-    if (fileUrl) {
+    if (file) {
+      const oldFileUrlRef = dbUser.fileUrl;
+      var oldRef = storage.refFromURL(oldFileUrlRef);
+      const storageRef = storage.ref();
+      const fileUID = uuidv4();
+      const fileRef = storageRef.child(fileUID);
+      await fileRef.put(file);
+      setFileUrl(await fileRef.getDownloadURL());
       updates.push(updateProfileFileUrl(fileUrl, currentUser.uid));
     }
-
-    const oldFileUrlRef = dbUser.fileUrl;
-    var oldRef = storage.refFromURL(oldFileUrlRef);
 
     Promise.all(updates)
       .then(() => {
@@ -125,16 +126,18 @@ export default function VolunteerUpdateProfile() {
       });
 
     // Delete the file
-    oldRef
-      .delete()
-      .then(() => {
-        // File deleted successfully
-        console.log("successfully deleted " + oldFileUrlRef);
-      })
-      .catch((error) => {
-        // Uh-oh, an error occurred!
-        console.log("unable to delete " + oldFileUrlRef);
-      });
+    if (oldRef) {
+      oldRef
+        .delete()
+        .then(() => {
+          // File deleted successfully
+          console.log("successfully deleted " + oldRef);
+        })
+        .catch((error) => {
+          // Uh-oh, an error occurred!
+          console.log("unable to delete " + oldRef);
+        });
+    }
   }
 
   function openModal() {
